Minor correction

+++ docs/literate/06_multimode_HG_sorter.jl
@@ -72,7 +72,7 @@ px = py = 3*w0  # Pitch between beams
 in_pos = Modes.TriangleLayout(n_groups, px, py, Shift2D(-2*px, 0) ∘ Rot2D(θ0))
 n_modes = length(in_pos)
 
-input_modes = generate_mode_stack(in_pos, ns, ds, Gaussian(Float32(w0)))
+input_modes = generate_mode_stack(in_pos, ns, ds, Gaussian(w0))
 u0 = ScalarField(input_modes, ds, λ)
 u0 = cu(u0)  #src
 #nb u0 = cu(u0)  # comment if you don't have CUDA
@@ -84,7 +84,7 @@ normalize_power!(u0, 1);
 # waist (wf = 8.0) than the input Gaussians. The modes are also rotated by 45°.
 
 wf = 8.0  # Output mode waist (HG modes need larger support)
-output_modes = generate_mode_stack(ns, ds, hermite_gaussian_groups(Float32(wf), n_groups);
+output_modes = generate_mode_stack(ns, ds, hermite_gaussian_groups(wf, n_groups);
                                    t = Rot2D(θ0))
 vf = ScalarField(output_modes, ds, λ)
 vf = cu(vf)  #src
